Working add Author, add article , get articles

+++ src/main/java/com/writerHub/practice/Controller/ArticleController.java
@@ -1,8 +1,14 @@
 package com.writerHub.practice.Controller;
 
 
-import com.writerHub.practice.models.Article;
+import com.auth0.jwt.exceptions.JWTVerificationException;
+import com.auth0.jwt.interfaces.DecodedJWT;
+import com.writerHub.practice.Util.JWTUtil;
+import com.writerHub.practice.models.*;
 import com.writerHub.practice.service.ArticleService;
+import com.writerHub.practice.service.AuthorService;
+import com.writerHub.practice.service.CompanyService;
+import com.writerHub.practice.service.WriterHubUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +22,13 @@ import java.util.List;
 public class ArticleController {
     @Autowired
     private ArticleService articleService;
+    @Autowired
+    private WriterHubUserService writerHubUserService;
+
+    @Autowired
+    private CompanyService companyService;
+    @Autowired
+    private AuthorService authorService;
 
     @GetMapping("/articles")
     public ResponseEntity<?> getAllArticles(){
@@ -23,19 +36,44 @@ public class ArticleController {
         return  ResponseEntity.ok().body(articleList);
     }
 
-    @PostMapping("/article")
-    public ResponseEntity<?> saveArticle(HttpServletRequest request, @RequestBody Article article){
-        Article savedArticle = articleService.saveArticle(article);
-        return new ResponseEntity<>(savedArticle, HttpStatus.CREATED);
+    @GetMapping("/author/{authorId}/articles")
+    public ResponseEntity<?> getAllArticlesByAuthor(@PathVariable Long authorId){
+        List<Article> articleList = articleService.getArticleByAuthorId(authorId);
+        return  ResponseEntity.ok().body(articleList);
     }
 
-    @DeleteMapping("/article/{id}")
-    public ResponseEntity<?> deleteArticle(@PathVariable Long id){
-        articleService.deleteArticle(id);
-        return ResponseEntity.ok().build();
+    @PostMapping("/author/{authorId}/article")
+    public ResponseEntity<?> saveArticle(@RequestBody Article article,@PathVariable Long authorId){
+        AuthenticationError error = new AuthenticationError();
+        try {
+            Author author = new Author(authorId);
+            article.setAuthor(author);
+            Article savedArticle = articleService.saveArticle(article);
+            return new ResponseEntity<>(savedArticle, HttpStatus.CREATED);
+        }
+        catch (Exception exception){
+            error.setStatusCode(HttpStatus.INTERNAL_SERVER_ERROR);
+            error.setMessage(exception.getMessage());
+        }
+        return new ResponseEntity<>(error,error.getStatusCode());
     }
 
-    @GetMapping("/article/{id}")
+    @DeleteMapping("author/{authorId}/article/{id}")
+    public ResponseEntity<?> deleteArticle(@PathVariable Long id,@PathVariable Long authorId){
+        Article article = articleService.getArticle(id);
+        if(article != null && article.getAuthor().getId() == authorId){
+            articleService.deleteArticle(id);
+            return ResponseEntity.ok().build();
+        }
+        else{
+            AuthenticationError error = new AuthenticationError();
+            error.setStatusCode(HttpStatus.UNAUTHORIZED);
+            error.setMessage("This article is not belongs to you.");
+            return new ResponseEntity<>(error,error.getStatusCode());
+        }
+    }
+
+    @GetMapping("author/{authorId}/article/{id}")
     public ResponseEntity<?> getArticle(@PathVariable Long id){
         Article article = articleService.getArticle(id);
         return ResponseEntity.ok().body(article);
